Fill branch delay slots
Remember the point where you could insert the branch the first time.
At the end the branch get scheduled. We know where it could be the
first time, so shift the instruction at max to this position in the
instruction stream. Additionally watch out for possible write
conflicts (currently only a save approach) and thread switches.

Signed-off-by: Jonas Pfeil <pfeiljonas@gmx.de>