Align the header of ralloc
The header of ralloc needs to be aligned, because otherwise the
compiler will make incorrect assumptions about the alignment leading
to degraded performance or aligment faults. On the raspberry pi
without aligment this creates a SIGBUS signal, leading to termination
of the program.

Also casting a not properly aligned pointer is undefined behaviour
in C.

This patch is not perfect for MSVC, as maybe in the future the
aligment for the most demanding data type might change to more than 8.

Signed-off-by: Jonas Pfeil <pfeiljonas@gmx.de>